Migrate navigation functions to controller.js
