+++ src/index.js
@@ -3,19 +3,19 @@ import './style.css';
 import {makeWebpage, makeCard} from './webpage/page.js';
 import makeButton from './webpage/buttons.js';
 import makeForm from './webpage/form.js';
+import makeController from './webpage/controller.js'
 
 const content = () => {
   const content = document.getElementById('content');
-  const overlay = document.createElement('div');
-  overlay.setAttribute('id', 'overlay');
-  // overlay.classList.add('active');
   
   const page = makeWebpage();
+  const form = makeForm();
+
   const header = page.header;
   const sidebar = page.sidebar;
   const main = page.main;
   const footer = page.footer;
-  const form = makeForm();
+  const overlay = page.overlay;
   
   content.append(header);
   content.append(sidebar);
@@ -24,95 +24,39 @@ const content = () => {
   content.append(form);
   content.append(overlay);
 
-  const openFormBtns = document.querySelectorAll('#add-task');
-  const closeFormBtns = document.querySelectorAll('#cancel-task');
-  // form.classList.add('active');
-
-  openFormBtns.forEach(button => {
-    button.addEventListener('click', () => {
-      const modal = document.querySelector('#add-task');
-      openForm(form, modal, overlay)
-    })
-  });
-
-  overlay.addEventListener('click', () => {
-    const modals = document.querySelectorAll('#cancel-task');
-    modals.forEach(modal => {
-      closeForm(form, modal, overlay);
-    })
-  })
-
-  closeFormBtns.forEach(button => {
-    button.addEventListener('click', () => {
-      const modal = document.querySelector('#cancel-task');
-      closeForm(form, modal, overlay)
-    })
-  })
+  const controller = makeController();
 };
 
-const openForm = (form, modal, overlay) => {
-  if (modal == null) return
-  modal.classList.add('active')
-  form.classList.add('active')
-  overlay.classList.add('active')
-};
+// const navigate = () => {
+//   const form = document.getElementById('form-container');
+//   const overlay = document.getElementById('overlay');
 
-const closeForm = (form, modal, overlay) => {
-  if (modal == null) return
-  modal.classList.remove('active')
-  form.classList.remove('active')
-  overlay.classList.remove('active')
-};
+//   const openFormBtns = document.querySelector('#add-task');
+//   const closeFormBtns = document.querySelector('#cancel-task');
 
-// const content = () => {
-//   const content = document.getElementById('content');
-//   const page = makeWebpage();
-  
-//   const header = page.header;
-//   const sidebar = page.sidebar;
-//   const main = page.main;
-//   const footer = page.footer;
-  
-//   content.append(header);
-//   content.append(sidebar);
-//   content.append(main);
-//   content.append(footer);
-  
-//   //const card = makeCard();
+//   openFormBtns.addEventListener('click', () => {
+//     openForm(form, overlay)
+//   });
 
-//   // main.append(makeCard());
-//   // main.append(makeCard());
-//   //const btnAddTask = makeButton('add-task', 'Add');
-//   //console.log(btnAddTask.id);
-//   //main.append(btnAddTask)
-//   //const addListener = addTaskListener(btnAddTask);
-//   const form = makeForm();
-//   main.append(form);
+//   overlay.addEventListener('click', () => {
+//     closeForm(form, overlay);
+//   });
 
-
-//   const nav = navigate();
+//   closeFormBtns.addEventListener('click', () => {
+//     closeForm(form, overlay)
+//   });
 // };
 
-// const addTaskListener = (btnPressed) => {
-//   const btn = document.querySelector(`#${btnPressed.id}`);
-//   btn.addEventListener('click', () => {
-//     main.append(makeCard());
-//   });
-// }
+// const openForm = (form, overlay) => {
+//   if (form == null) return
+//   form.classList.add('active')
+//   overlay.classList.add('active')
+// };
 
-// TODO!!
-// const navigate = () => {
-//   const main = document.getElementById('main');
-//   const addBtn = document.querySelector('#add-task');
-//   const cancelBtn = document.querySelector('#cancel-task');
-//   const createBtn = document.querySelector('#create-task');
-  
-//   addBtn.addEventListener('click', () => {
-//     //main.append(form);
-//   });
-//   cancelBtn.addEventListener('click', () => {
-//     //main.remove(form);
-//   });
+// const closeForm = (form, overlay) => {
+//   if (form == null) return
+//   form.classList.remove('active')
+//   overlay.classList.remove('active')
 // };
 
 content();
